+++ src/tests/intg/Makefile.am
@@ -1,6 +1,7 @@
 dist_noinst_DATA = \
     config.py.m4 \
     util.py \
+    sssd_nss.py \
     sssd_id.py \
     sssd_ldb.py \
     sssd_netgroup.py \
